requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
aiomysql==0.2.0
pydantic==2.5.2
beautifulsoup4==4.12.2
requests==2.31.0
aiohttp==3.9.1
python-dotenv==1.0.0
schedule==1.2.1
nltk==3.8.1
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
transformers==4.35.2
torch==2.1.1 